# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{organizationId}/whatsapp/integration:
    get:
      operationId: getWhatsAppIntegration
      tags: [whatsapp-integration]
      summary: Get WhatsApp integration settings (owner/admin only)
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Integration settings without credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WhatsAppIntegration" }
    patch:
      operationId: updateWhatsAppIntegration
      tags: [whatsapp-integration]
      summary: Update WhatsApp integration settings (owner/admin only)
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WhatsAppIntegrationUpdate" }
      responses:
        "200":
          description: Updated integration settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WhatsAppIntegration" }
...
  /organizations/{organizationId}/whatsapp/sync:
    post:
      operationId: syncWhatsAppResources
      tags: [whatsapp-integration]
      summary: Synchronize configured WABA resources (owner/admin only)
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Synchronization result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WhatsAppSyncResult" }
  /organizations/{organizationId}/whatsapp/wabas:
    get:
      operationId: listWhatsAppWabas
      tags: [whatsapp-integration]
      summary: List synchronized WABAs (owner/admin only)
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Synchronized WABAs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WhatsAppWaba" }
...
components:
  parameters:
    OrganizationIdParam:
      name: organizationId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    WhatsAppProviderMode:
      type: string
      enum: [mock, real]
    WhatsAppIntegration:
      type: object
      required: [organizationId, mode, status, health, webhookVerificationConfigured, webhookSignatureConfigured, createdAt, updatedAt]
      properties:
        organizationId: { type: integer }
        mode: { $ref: "#/components/schemas/WhatsAppProviderMode" }
        connectorAccountId: { type: ["string", "null"], description: Non-secret connector identity, available only after real-mode verification. }
        configuredWabaExternalId: { type: ["string", "null"] }
        status: { type: string, enum: [unconfigured, configured, syncing, healthy, error] }
        health: { type: string, enum: [unknown, healthy, unhealthy] }
        lastHealthAt: { type: ["string", "null"], format: date-time }
        lastSyncAt: { type: ["string", "null"], format: date-time }
        lastErrorAt: { type: ["string", "null"], format: date-time }
        lastError: { type: ["string", "null"] }
        webhookVerificationConfigured: { type: boolean }
        webhookSignatureConfigured: { type: boolean }
        realWorkspaceLimit: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    WhatsAppIntegrationUpdate:
      type: object
      properties:
        mode: { $ref: "#/components/schemas/WhatsAppProviderMode" }
        configuredWabaExternalId: { type: ["string", "null"], maxLength: 128 }
...
    WhatsAppSyncResult:
      type: object
      required: [wabas, phoneNumbers, templates, syncedAt]
      properties:
        wabas: { type: integer }
        phoneNumbers: { type: integer }
        templates: { type: integer }
        syncedAt: { type: string, format: date-time }
...
    WhatsAppWaba:
      type: object
      required: [id, externalId, displayName, provider, createdAt, updatedAt]
      properties:
        id: { type: integer }
        externalId: { type: string }
        displayName: { type: string }
        provider: { type: string }
        providerStatus: { type: ["string", "null"] }
        lastSyncedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...